refactor: unify more corecrypto e2ei impls in a common module
# Conflicts:
#	crypto-ffi/src/generic/mod.rs
#	crypto-ffi/src/wasm/mod.rs
